requirements.txt
tqdm
ifaddr
pillow
kivy
platformdirs
